% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_occ_uncertain_data.R
\name{create_occ_uncertain_data}
\alias{create_occ_uncertain_data}
\title{Create Occurrence Data with Uncertainty}
\usage{
create_occ_uncertain_data(
  occs_df,
  lat_col = "latitude",
  lon_col = "longitude",
  coord_uncert,
  na_action = "NA as NA"
)
}
\arguments{
\item{occs_df}{A \code{data.frame} of occurrence locations that includes
\emph{at least these four columns} - latitude, longitude, latitude uncertainty and longitude uncertainty in degrees.}

\item{lat_col}{Name of column of latitude dbl values. Caps sensitive.}

\item{lon_col}{Name of column of longitude dbl values. Caps sensitive.}

\item{coord_uncert}{Name of column of coordinate uncertainty values,
in meters. Case sensitive.}

\item{na_action}{Sets action to take regarding occurrences with coordinate
uncertainty values of NA. Case sensitive}

\item{lat_uncertainty}{Name of column of latitude uncertainty in degree values. Caps sensitive.}

\item{lon_uncertainty}{Name of column of longitude uncertainty in degree values. Caps sensitive.}

\item{taxa_col}{Name of column of taxa (species) values. Caps sensitive.}
}
\description{
Function to create occurrence data sets with spatial uncertainty incorporated
into point positions.
}
\details{
\strong{Input} as a \code{dataframe} must have the following columns:
\tabular{ccccc}{
[,1] \tab ddlat \tab numeric, latitude (in decimal degrees)\cr
[,2] \tab ddlon \tab numeric, longitude (in decimal degrees)\cr
[,3] \tab ddlat unc \tab numeric, longitude uncertainty (in decimal degrees)\cr
[,4] \tab ddlon unc \tab numeric, longitude uncertainty (in decimal degrees)\cr
[,5] \tab tax \tab character or factor, taxa names\cr}

There can be additional columns in the data frame, which will be included
in the output data frames, but otherwise ignored.
}
