% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meters_to_decdeg.R
\name{meters_to_decdeg}
\alias{meters_to_decdeg}
\title{Convert from meters to degrees correcting for global position}
\usage{
meters_to_decdeg(
  occs_df,
  lat_col = "latitude",
  lon_col = "longitude",
  coord_uncert,
  na_action = "NA as NA"
)
}
\arguments{
\item{occs_df}{A \code{data.frame} of occurrence locations that includes
\emph{at least these three columns} - latitude, longitude, and coordinate
uncertainty values (in meters to be converted to decimal degrees).}

\item{lat_col}{Name of column of latitude values. Case sensitive.}

\item{lon_col}{Name of column of longitude values. Case sensitive.}

\item{coord_uncert}{Name of column of coordinate uncertainty values,
in meters. Case sensitive.}

\item{na_action}{Sets action to take regarding occurrences with coordinate
uncertainty values of NA. Case sensitive}
}
\value{
dist_dd A \code{data.frame} of longitude and latitude
   (in that column order) distances in units of degree decimal.
}
\description{
\code{meters_to_decdeg} converts from meters to degrees at a specified
position on the globe. The use case this function was developed for was to
calculate occurrence point uncertainty values, which are usually reported in
meters, as degrees.

The formula for converting from meters to decimal degrees is in part
based on information from the ESRI ArcUser magazine at this site
\url{https://www.esri.com/news/arcuser/0400/wdside.html}
}
