% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_geo_range.R
\name{random_geo_range}
\alias{random_geo_range}
\title{Generate random geographic range of species area of occupancy (AOO) and extent of occurrence (EOO), from latitude and longitude coordinates accounting uncertainty values}
\usage{
random_geo_range(n_length, occs_df, lat_col = "latitude",
  lon_col = "longitude", lat_uncertainty = "lat_uncertainty",
  lon_uncertainty = "lon_uncertainty", taxa_col = "species", ...)
}
\arguments{
\item{n_length}{Number of loops to be run.}

\item{occs_df}{A \code{data.frame} of occurrence locations that includes
\emph{at least these five columns} - length of loops to be run. latitude, longitude, latitude uncertainty and longitude uncertainty in degrees.}

\item{lat_col}{Name of column of latitude values. Caps sensitive.}

\item{lon_col}{Name of column of longitude values. Caps sensitive.}

\item{lat_uncertainty}{Name of column of latitude uncertainty in degree values. Caps sensitive.}

\item{lon_uncertainty}{Name of column of longitude uncertainty in degree values. Caps sensitive.}

\item{taxa_col}{Name of column of taxa (species) values. Caps sensitive.}

\item{lon_random}{latitude random deviates of the interval from min to max. Caps sensitive.}

\item{lon_random}{longitude random deviates of the interval from min to max. Caps sensitive.}
}
\description{
\code{random_geo_range} Generate random geographic range of AOO and EOO, from latitude and longitude coordinates accounting uncertainty values. The use case this function was developed for was to generate several AOO and EOO.
}
