% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leopardus_wiedii_gbif.R
\docType{data}
\name{Leopardus_wiedii_gbif}
\alias{Leopardus_wiedii_gbif}
\title{Occurence data for \emph{Leopardus wiedii} aquired from GBIF}
\format{An object of class \code{data.frame} with 229 rows and 135 columns.}
\source{
\url{http://www.gbif.org}
}
\usage{
Leopardus_wiedii_gbif
}
\description{
Give some more description of this data set, including the
associated GBIF citation.
}
\keyword{datasets}
