% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leopardus_wiedii_gbif.R
\docType{data}
\name{Leopardus_wiedii_gbif}
\alias{Leopardus_wiedii_gbif}
\title{Occurence data for \emph{Leopardus wiedii} acquired from GBIF}
\format{
An object of class \code{data.frame} with 229 rows and 4 columns.
}
\source{
\url{https://www.gbif.org}
}
\usage{
Leopardus_wiedii_gbif
}
\description{
The margay (Leopardus wiedii, Schinz 1821) is a small Neotropical cat that is native with distribution from South America into Northeastern Mexico.  Citation - GBIF.org (January 2023) GBIF Occurrence Download http://api.gbif.org/v1/occurrence/download/request/0066556-200221144449610.zip
}
\details{
DOI 10.15468/dl.ljhrm9
}
\keyword{datasets}
